% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorbing_states.R
\name{absorbing_states}
\alias{absorbing_states}
\title{Identify absorbing states of a Markov chain}
\usage{
absorbing_states(P, mode = "binarr")
}
\arguments{
\item{P}{The transition probability matrix.}

\item{mode}{Can take value "binarr" (producing a binary array) or "indices" (returning the absorbent state indices).}
}
\value{
A binary array (if mode == "binarr") or an array of indices (if mode == "indices").
}
\description{
This function takes the transition probability matrix as input and finds absorbent states.
}
\examples{
mtx = diag(1, 3, 3)
absorbing_states(mtx, "binarr")
absorbing_states(mtx, "indices")

}
