% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\title{This function takes the transition probability matrix as input and finds states accessible from each state.}
\usage{
access(P)
}
\arguments{
\item{P}{The transition probability matrix.}
}
\value{
A matrix of binary values where the i, j th entry indicates whether the jth state is accessible from the ith one.
}
\description{
This function takes the transition probability matrix as input and finds states accessible from each state.
}
\examples{
mtx = diag(1, 3, 3)
access(mtx)

mtx2 = matrix(1/3, 3, 3)
access(mtx2)

}
