% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_classes.R
\name{classify_classes}
\alias{classify_classes}
\title{Classifies equivalent classes as recurrent and transient.}
\usage{
classify_classes(P, mode = "bin_arr")
}
\arguments{
\item{P}{The transition probability matrix.}

\item{mode}{"binarr" (returns binary vector with ith entry indicating whether the ith class is recurrent) or "visual" (returns readable dataframe with class elements and classification)}
}
\value{
A binary array or dataframe
}
\description{
This function takes the transition probability matrix as input and returns the classification of equivalent classes.
}
\examples{
mtx = diag(1, 3, 3)
classify_classes(mtx, "visual")

mtx2 <- matrix(1/3, 3, 3)
classify_classes(mtx2)

}
