% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_states.R
\name{classify_states}
\alias{classify_states}
\title{Function to classify states as recurrent or transient}
\usage{
classify_states(P, mode = "bin_arr")
}
\arguments{
\item{P}{The transition probability matrix.}

\item{mode}{Ouptut format: Either "binarr" or "visual"}
}
\value{
A matrix of binary values or dataframe
}
\description{
This function takes the transition probability matrix as input and returns whether it is recurrent (1) or transient (0)
}
\examples{
mtx <- matrix(c(0, 1, 0, 0, 0, 1, 0, 1, 0 ),3, 3, byrow =TRUE)
classify_states(mtx, "visual")

mtx2 <- matrix(1/3, 3, 3)
classify_states(mtx2)

}
