% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_classes.R
\name{comm_classes}
\alias{comm_classes}
\title{Finds communicating classes in a stochastic process.}
\usage{
comm_classes(P, mode = "bin_arr")
}
\arguments{
\item{P}{The transition probability matrix.}

\item{mode}{"binarr" (returns binary array with i,j th entry indicating whether jth element is present in ith class) or "list" (with ith element of list being a vector of states present in ith class)}
}
\value{
A binary array or list of vectors of
}
\description{
This function takes the transition probability matrix as input and returns the communicating classes.
}
\examples{
mtx = diag(1, 3, 3)
comm_classes(mtx, "list")

mtx2 <- matrix(1/3, 3, 3)
comm_classes(mtx2)

}
