% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stochastic.R
\name{is_stochastic}
\alias{is_stochastic}
\title{Checks whether a matrix is stochastic.}
\usage{
is_stochastic(P)
}
\arguments{
\item{P}{The matrix.}
}
\value{
A boolean value representing whether it is stochastic.
}
\description{
This function takes any matrix as input and finds whether it is a valid transition probability matrix of a stochastic process.
}
\examples{
mtx = diag(1, 3, 3)
is_stochastic(mtx)

mtx2 <- matrix(1, 2, 3)
is_stochastic(mtx2)

}
