% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem_onehot.R
\name{mem_onehot}
\alias{mem_onehot}
\title{Helper function to create concatenated one hot vectorss}
\usage{
mem_onehot(idx, n)
}
\arguments{
\item{idx}{A numeric matrix containing the indices to convert to one hot}

\item{n}{The number of states in the MC}
}
\value{
A matrix with concatenated one hot vectors
}
\description{
Helper function to create concatenated one hot vectorss
}
\examples{
idx <- c(2,3)
mem_onehot(idx, 6) 

}
