% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachable.R
\name{reachable}
\alias{reachable}
\title{Checks whether a state is reachable from another state.}
\usage{
reachable(P, i, j)
}
\arguments{
\item{P}{The transition probability matrix.}

\item{i}{The original state.}

\item{j}{The target state.}
}
\value{
A boolean value representing whether it is reachable.
}
\description{
This function takes the transition probability matrix as input and finds whether the jth state is accessible/reachable from the ith state.
}
\examples{
mtx = diag(1, 3, 3)
reachable(mtx, 1, 2)

mtx2 <- matrix(1/3, 3, 3)
reachable(mtx2, 1, 2)

}
