% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_class_probs.R
\name{steady_class_probs}
\alias{steady_class_probs}
\title{Function to find steady state probabilities of entering into different communicating classes}
\usage{
steady_class_probs(P)
}
\arguments{
\item{P}{The transition probability matrix.}
}
\value{
A matrix (n_states * n_classes) of probabilities
}
\description{
This function takes the transition probability matrix as input and returns probabilities of ending up in each class from each state
}
\examples{
test <- matrix(c(0, 0.25, 0.5, 0.25, 0, 0,
0, 0.5, 0, 0.5, 0, 0,
0, 0, 0.25, 0, 0.75, 0,
0, 0.5, 0, 0.5, 0, 0,
0, 0, 0.75, 0, 0.25, 0,
0, 0.25, 0, 0.25, 0.5, 0
) , 6, 6, byrow = TRUE)

steady_class_probs(test)

}
