% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_recurrent.R
\name{steady_recurrent}
\alias{steady_recurrent}
\title{Outputs stable configuration for a recurrent Markov chain}
\usage{
steady_recurrent(P)
}
\arguments{
\item{P}{The transition probability matrix.}
}
\value{
The probability matrix of the states after t= infinity.
}
\description{
This function takes the transition probability matrix for a recurrent Markov chain as input and finds absorbent states.
}
\examples{
mtx = matrix(0.25, 4, 4)
steady_recurrent(mtx)

}
