% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_return_probs.R
\name{steady_return_probs}
\alias{steady_return_probs}
\title{Function to find steady state probabilities of returning to the initial state}
\usage{
steady_return_probs(P)
}
\arguments{
\item{P}{The transition probability matrix.}
}
\value{
A matrix (1 * n_states) of probabilities
}
\description{
This function takes the transition probability matrix as input and returns probabilities of returning to the same state at t = infinity.
}
\examples{
test <- matrix(c(0, 0.25, 0.5, 0.25, 0, 0,
0, 0.5, 0, 0.5, 0, 0,
0, 0, 0.25, 0, 0.75, 0,
0, 0.5, 0, 0.5, 0, 0,
0, 0, 0.75, 0, 0.25, 0,
0, 0.25, 0, 0.25, 0.5, 0
) , 6, 6, byrow = TRUE)

steady_return_probs(test)

}
