% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{steady_state}
\alias{steady_state}
\title{Function to find stationary matrix (probabilities in steady state)}
\usage{
steady_state(P)
}
\arguments{
\item{P}{The transition probability matrix.}
}
\value{
A stochastic matrix (n_states * n_states) of probabilities
}
\description{
This function takes the transition probability matrix as input and returns the probability matrix of ending up at each state at t = infinity.
}
\examples{
test <- matrix(c(0, 0.25, 0.5, 0.25, 0, 0,
0, 0.5, 0, 0.5, 0, 0,
0, 0, 0.25, 0, 0.75, 0,
0, 0.5, 0, 0.5, 0, 0,
0, 0, 0.75, 0, 0.25, 0,
0, 0.25, 0, 0.25, 0.5, 0
) , 6, 6, byrow = TRUE)

steady_state(test)

}
