% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submtx.R
\name{submtx}
\alias{submtx}
\title{Helper function to create a submatrix of a TPM consisting of specific states}
\usage{
submtx(P, idx)
}
\arguments{
\item{P}{The TPM}

\item{idx}{A numeric matrix containing the indices}
}
\value{
A matrix with only the specific states (generally for a recurrent class)
}
\description{
Helper function to create a submatrix of a TPM consisting of specific states
}
\examples{
idx <- c(2,3)
P <- matrix(c(1,0,0, 0, 1, 0, 0, 0, 1), 3, 3, byrow =TRUE)
submtx(P, idx) 

}
